+++ kern/sys/acpi/acpi.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 #include "../../include/limine.h"
 
 typedef struct {
@@ -203,4 +204,4 @@ typedef struct {
 int acpi_table_checksum(xsdt_t *t);
 void *find_acpi_table(char signature[4], rsdt_t *rsdt, xsdt_t *xsdt);
 void init_acpi(void);
-void pmt_delay(int us);
+void pmt_delay(size_t us);
